Changed service and component of list product to work with the new Product changes

+++ angular-frontend/src/app/list-product/list-product.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams, HttpHeaders} from "@angular/common/http";
-import {Observable} from "rxjs";
+import {Observable, throwError} from "rxjs";
 
 @Injectable({
   providedIn: 'root'
@@ -25,9 +25,13 @@ export class ListProductService {
     });
 
   }
-  deleteProduct(id:string):Observable<any> {
-    let httpParams = new HttpParams()
-      .set('id', id);
-    return this.http.delete(this.ROOT_URL, {params:httpParams});
+  deleteProduct(id:string | undefined):Observable<any> {
+    if (id != undefined) {
+      let httpParams = new HttpParams()
+        .set('id', id);
+      return this.http.delete(this.ROOT_URL, {params: httpParams});
+    }else {
+      return throwError("Cant find Id of the object ot delete id");
+    }
   }
 }
